code/requirements.txt
requests==2.22.0
Keras==2.2.5
numpy==1.17.1
Pillow==6.1.0
opencv-python==4.1.0.25
tensorflow=1.14.0
